+++ untitled2/src/Plot.java
@@ -1,7 +1,7 @@
 import java.awt.*;
 
 public class Plot extends Shapes {
-    public Plot(int x, int y, Color c) {
+    public Plot(double x, double y, Color c) {
         super.type = 0;
         super.colour = c;
         super.x[0] = x;
